requirements.txt
astroid==1.6.6
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.0.5
django-admin-view-permission==1.6
django-colorfield==0.1.14
django-cors-headers==2.3.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.6
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
oauthlib==2.1.0
Pillow==5.1.0
psycopg2==2.7.4
PyJWT==1.6.4
pylint==1.9.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==1.0.0
sendotp==0.0.4
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.22
whitenoise==3.3.1
wrapt==1.10.11
XlsxWriter==1.1.8
